\name{block_figure}
\alias{block_figure}
\title{Embed a plot in a floating figure block...}
\usage{block_figure(code, outdir=".", plot_width=4, plot_height=4,
    tex_width="0.5\linewidth", tex_height, dpi=300, position="htbp",
    caption, label, ..., col=2, envir=globalenv())}

\description{
  Embed a plot in a floating figure block Display all plots produced by
  the code.
}
\arguments{
  \item{col}{number of columns}
}
