\name{block_interweave}
\alias{block_interweave}
\title{Interweave code and output, as if you had executed at the command line...}
\usage{block_interweave(code, plot_width=4, plot_height=4,
    tex_width="0.5\linewidth", tex_height, dpi=300, ...,
    envir=globalenv())}

\description{
  Interweave code and output, as if you had executed at the command
  line
}
