\name{decumar}
\alias{decumar}
\title{Overwrite input with processed output...}
\usage{decumar(in_path, out_path=in_path, clean=FALSE)}

\description{
  Overwrite input with processed output
}
\arguments{
  \item{in_path}{path to input file}
  \item{out_path}{path where output should be saved.  Defaults to overwriting
input}
  \item{reset}{should result cache be emptied before processing?}
}
